Move the repository to the repository dir
